Added comments and tests.
Added comments to the GCRPolicy Struct, explaining the purpose and use of each value.
Added 2 preset values to GCRPolicy, which are MediumStrength and HighStrength. Mostly used for passwords to define strength.
Added a global function, verifyEmail, which will verify an email address in a string.

Finally, I added a series of tests to a new test class, EmailVerifyTests, to test the verifyEmail function to confirm its functionality.