Moved verifyEmail.
Verify Email function is now a static function within the GCRSwiftlyLogin class. This is to allow it to be used within that class.
Tests have been updated appropriately for the refactoring of verifyEmail and the personal email test has been updated to use an email address that can be public.